Backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
groq
langchain
langchain-community
langchain-text-splitters
pypdf
supabase
langchain-groq
langgraph
langchainhub
transformers
torch
torchvision
Pillow